Fix error handling in builder Prepare()
In order for Prepare() to be considered "successful", the returned
error must be nil. However, after calling packer.MultiErrorAppend,
the errs variable is not nil anymore, always resulting in a failed
image build.

This patch follows the pattern found in other builder plugins to test
whether the packer.MultiError contains any errors. If so, the errors
are returned, if not, a literal nil is returned.