Use sources from /vagrant as build source
The working directory is supposed to contain the git checkout of
packer-builder-arm-image. Instead of cloning again in the vagrant
provisioning, just copy it over from /vagrant.

This allows quickly testing of source changes without committing
and pushing to master and to even do integration tests on branches
other than master.