Read image output name from packer output.
Instead of relying on the "output-arm-image" name prefix, the
output of packer is parsed to determine the actual path of the
resulting image.
This allows for setting the build name variable using the "name"
attribute in the JSON configuration (resulting in output-BUILDNAME)
or completely overriding the output path using the "output_directory"
attribute.
Both variants will work and allow configuration of the resulting
output file name even in Vagrant.