Fix 'file already closed' error for compressed images
When unpacking compressed images that are not compressed as zip archives
(e.g. gzip, xz) the image would unpack nearly all of the way and then
would fail with an error like: 'read |0: file already closed' as is
documented in GitHub issue
[61](https://github.com/solo-io/packer-plugin-arm-image/issues/61).

As the decompression is offloaded to an external tool for these
compression types and the data is streamed through a reader and writer
in order to allow reporting on progress, we ensure that the exec.Cmd
instance for the external decompression tool is not torn down until the
associated reader is closed. The
[docs](https://pkg.go.dev/os/exec#Cmd.StdoutPipe) for Cmd.StdoutPipe
explain this in more detail.