Add image_arch and qemu_required options
The inclusion of image_arch improves detecting when qemu is required (#145).
The image architecture is also used to intelligently determine which flavor
of qemu is required, if any.
Additionally, the qemu_required option had been added which forces the use
of qemu even when it's not stricly required by qemu_args, qemu_binary, or
image_arch.  This aims to help in build situations where the user wants
reproducible emulation or when QEMU_* environment variables need to be respected.
The default values for both of these options have been chosen to prevent
any breaking changes to existing configurations.
One example configuration has been updated to enable native image building
should the build environment support it.